requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development
black>=23.10.0
isort>=5.12.0
mypy>=1.7.0
pylint>=3.0.0

# LLM Integration
openai>=1.3.0
transformers>=4.35.0

# Storage
aiosqlite>=0.19.0

# Sourcegraph Integration
requests>=2.31.0
